Поиск в журнале перенесен в базовый класс журнала.
Он нужен во всех журналах, даже реализованных на JournalViewModelBase при этом в текущей реализации его все равно ни на что другое не заменить.
