+++ QS.Project/Project.Journal/EntitiesJournalViewModelBase.cs
@@ -47,13 +47,6 @@ namespace QS.Project.Journal
 			this.commonServices = commonServices ?? throw new ArgumentNullException(nameof(commonServices));
 			UseSlider = true;
 			EntityConfigs = new Dictionary<Type, JournalEntityConfig<TNode>>();
-			Search.OnSearch += Search_OnSearch;
-			searchHelper = new SearchHelper(Search);
-		}
-
-		void Search_OnSearch(object sender, EventArgs e)
-		{
-			Refresh();
 		}
 
 		void FilterViewModel_OnFiltered(object sender, EventArgs e)
@@ -127,22 +120,6 @@ namespace QS.Project.Journal
 		}
 
 		#endregion Ordering
-
-		#region Search
-
-		private readonly SearchHelper searchHelper;
-
-		protected ICriterion GetSearchCriterion(params Expression<Func<object>>[] aliasPropertiesExpr)
-		{
-			return searchHelper.GetSearchCriterion(aliasPropertiesExpr);
-		}
-
-		protected ICriterion GetSearchCriterion<TEntity>(params Expression<Func<TEntity, object>>[] propertiesExpr)
-		{
-			return searchHelper.GetSearchCriterion(propertiesExpr);
-		}
-
-		#endregion Search
 
 		#region Entity load configuration
